Move TSMixer block to shared blocks
