+++ pytorch_forecasting/layers/_blocks/__init__.py
@@ -3,9 +3,11 @@ from pytorch_forecasting.layers._blocks._scinet_block import SCIBlock
 from pytorch_forecasting.layers._blocks._softs_block import (
     STADModule,
 )
+from pytorch_forecasting.layers._blocks._tsmixer_block import TSMixerBlock
 
 __all__ = [
     "ResidualBlock",
     "SCIBlock",
     "STADModule",
+    "TSMixerBlock",
 ]
